Remove complete -F _longopt mv complete -F _root_command gksudo complete -F _command nice complete -F _longopt tr complete -F _longopt head complete -F _service /etc/init.d/rsync complete -F _longopt sha256sum complete -F _longopt uniq complete -F _command else complete -F _root_command gksu complete -F _service /etc/init.d/unattended-upgrades complete -F _longopt ptx complete -F _longopt mkfifo complete -F _longopt tee complete -F _known_hosts traceroute complete -j -P '"%' -S '"' fg complete -F _longopt who complete -F _service /etc/init.d/x11-common complete -o default -F __start_kubectl kubectl complete -F _longopt grep complete -F _longopt objdump complete -F _longopt less complete -F _longopt sha1sum complete -F _longopt cut complete -F _longopt mknod complete -F _command padsp complete -F _service /etc/init.d/plymouth complete -F _service /etc/init.d/console-setup.sh complete -F _longopt bison complete -F _command nohup complete -F _longopt od complete -a unalias complete -F _service /etc/init.d/kmod complete -u groups complete -F _longopt texindex complete -F _known_hosts telnet complete -F _completion_loader -D complete -F _command vsound complete -F _export export complete -c which complete -F _longopt m4 complete -F _longopt cp complete -F _longopt base64 complete -F _longopt strip complete -o bashdefault -o default -o nospace -F __git_wrap__gitk_main gitk complete -F _longopt split complete -F _service /etc/init.d/screen-cleanup complete -v readonly complete -o nospace -F _cd cd complete -F _known_hosts showmount complete -F _longopt tac complete -F _longopt fold complete -F _known_hosts mtr complete -F _known_hosts fping complete -F _longopt env complete -F _user_at_host finger complete -c type complete -o dirnames -o filenames -F _apport-unpack apport-unpack complete -F _service /etc/init.d/apport complete -F _known_hosts ssh-installkeys complete -F _root_command kdesudo complete -F _longopt expand complete -o bashdefault -o default -o nospace -F __git_wrap__git_main git complete -F _longopt ln complete -F _command aoss complete -u w complete -F _service /etc/init.d/ufw complete -F _longopt ld complete -F _service /etc/init.d/apparmor complete -F _longopt enscript complete -F _longopt irb complete -F _command xargs complete -j -P '"%' -S '"' jobs complete -F _service service complete -F _longopt tail complete -F _longopt unexpand complete -F _longopt netstat complete -o dirnames -o filenames -F _apport-cli apport-cli complete -F _longopt ls complete -F _command tsocks complete -F _longopt diff complete -o nospace -F _cd pushd complete -v unset complete -F _longopt csplit complete -F _known_hosts rsh complete -F _command exec complete -F _longopt shar complete -F _longopt vdir complete -F _longopt sum complete -F _longopt nm complete -F _longopt nl complete -F _user_at_host ytalk complete -F _service /etc/init.d/plymouth-log complete -j -P '"%' -S '"' disown complete -o dirnames -o filenames -F _apport-bug apport-bug complete -F _longopt bash complete -F _service /etc/init.d/procps complete -u sux complete -F _longopt md5sum complete -A stopped -P '"%' -S '"' bg complete -F _longopt objcopy complete -F _longopt bc complete -F _longopt paste complete -F _longopt dir complete -b builtin complete -F _longopt a2ps complete -F _root_command really complete -F _longopt shasum complete -F _command ltrace complete -F _service /etc/init.d/hwclock.sh complete -F _service /etc/init.d/irqbalance complete -F _known_hosts traceroute6 complete -F _longopt date complete -F _longopt cat complete -F _longopt readelf complete -F _longopt awk complete -F _known_hosts dig complete -F _user_at_host talk complete -F _longopt sha512sum complete -F _longopt seq complete -F _longopt mkdir complete -F _longopt sha224sum complete -A helptopic help complete -F _minimal '' complete -F _service /etc/init.d/keyboard-setup.sh complete -F _longopt df complete -F _command eval complete -A setopt set complete -F _longopt chroot complete -F _longopt sort complete -F _command do complete -F _service /etc/init.d/dbus complete -F _longopt pr complete -F _longopt colordiff complete -F _longopt du complete -F _longopt fmt complete -F _longopt sed complete -F _longopt gperf complete -F _command time complete -F _service /etc/init.d/udev complete -F _longopt wc complete -F _root_command fakeroot complete -u slay complete -F _longopt grub complete -F _service /etc/init.d/cron complete -F _longopt rmdir complete -F _service /etc/init.d/uuidd complete -A shopt shopt complete -F _apport-collect apport-collect complete -o dirnames -o filenames -F _apport-bug ubuntu-bug complete -F _known_hosts ftp complete -F _longopt units complete -F _longopt uname complete -F _longopt touch complete -F _longopt ldd complete -F _command then complete -F _known_hosts rlogin complete -F _command command complete -F _longopt sha384sum complete -F _known_hosts fping6 complete -F _longopt rm flag when scope is

+++ entrypoint.sh
@@ -3,6 +3,7 @@ set -euo pipefail
 
 # Define colors
 RED='\033[0;31m'
+YELLOW='\033[0;33m'
 GREEN='\033[0;32m'
 NC='\033[0m' # No Color (reset)
 
@@ -96,22 +97,29 @@ jq 'with_entries(.value |= {value: .})' "$params_path" > params.json
 # Authenticate with Azure using the service principal
 echo "Authorizing to Azure using service principal..."
 if ! az login --service-principal -u "$azure_client_id" -p "$azure_client_secret" -t "$azure_tenant_id"; then
-  echo "Authentication failed. Please check the Azure credentials and refer to provisioner documentation."
+  echo -e "${RED}Authentication failed. Please check the Azure credentials and refer to provisioner documentation.${NC}"
   exit 1
 fi
 az account set --subscription "$azure_subscription_id"
-echo "${GREEN}Authentication successful.${NC}"
+echo -e "${GREEN}Authentication successful.${NC}\n"
 
 az_flags=""
 # Set command and flag settings based on scope
 case "$scope" in
   group)
-    echo "Targeting resource group $resource_group"
+    echo "Targeting resource group $resource_group\n"
     az_flags+=" --resource-group $resource_group"
+
+    if [ "$complete" = "true" ] && [ "$MASSDRIVER_DEPLOYMENT_ACTION" != "decommission" ]; then
+      az_flags+=" --mode Complete"
+    fi
     ;;
   sub)
-    echo "Targeting subscription $azure_subscription_id"
-    az_flags+=" --location $location"
+    echo "Targeting subscription $azure_subscription_id\n"
+    if [ "$MASSDRIVER_DEPLOYMENT_ACTION" != "decommission" ]; then
+      az_flags+=" --location $location"
+    fi
+    
     ;;
   *)
     echo -e "${RED}Error: Unsupported scope '$scope'. Expected 'group' or 'sub'.${NC}"
@@ -119,10 +127,7 @@ case "$scope" in
     ;;
 esac
 
-# Extract flags from provisioner config
-if [ "$complete" = "true" ]; then
-  az_flags+=" --mode Complete"
-fi
+
 
 deployment_name="${name_prefix}-${MASSDRIVER_STEP_PATH}"
 
@@ -132,7 +137,7 @@ case "$MASSDRIVER_DEPLOYMENT_ACTION" in
   plan)
     evaluate_checkov
     echo "Executing plan..."
-    az deployment $scope what-if $az_flags --name $deployment_name --template-file template.bicep --parameters @params.json --parameters @connections.json
+    az deployment $scope what-if $az_flags --name "$deployment_name" --template-file template.bicep --parameters @params.json --parameters @connections.json
     ;;
 
   provision)
@@ -156,7 +161,7 @@ case "$MASSDRIVER_DEPLOYMENT_ACTION" in
     fi
 
     echo "Deploying bundle"
-    az deployment $scope create $az_flags --name $deployment_name --template-file template.bicep --parameters @params.json --parameters @connections.json | tee outputs.json
+    az deployment $scope create $az_flags --name "$deployment_name" --template-file template.bicep --parameters @params.json --parameters @connections.json | tee outputs.json
 
     jq -s '{params:.[0],connections:.[1],envs:.[2],secrets:.[3],outputs:.[4].properties.outputs}' "$params_path" "$connections_path" "$envs_path" "$secrets_path" outputs.json > artifact_inputs.json
     for artifact_file in artifact_*.jq; do
@@ -168,13 +173,17 @@ case "$MASSDRIVER_DEPLOYMENT_ACTION" in
     ;;
 
   decommission)
-    echo "Decommissioning resources..."
-    touch empty.bicep
-    az deployment $scope create $az_flags --name $deployment_name --template-file empty.bicep
-    rm empty.bicep
+    if [ "$scope" = "group" ] && [ "$complete" = "true" ]; then
+      echo "Decommissioning resources..."
+      touch empty.bicep
+      az deployment $scope create $az_flags --mode Complete --name "$deployment_name" --template-file empty.bicep
+      rm empty.bicep
+    else
+        echo -e "${YELLOW}Your bundle configuration prevents resource deletion during decommission. Be sure to delete bundle resources manually.${NC}\n"
+    fi
 
     echo "Deleting deployment group $deployment_name"
-    az deployment $scope delete --name "$deployment_name"
+    az deployment $scope delete $az_flags --name "$deployment_name"
 
     if [ "$scope" = "group" ] && [ "$delete_resource_group" = "true" ]; then
       echo "Deleting resource group $resource_group in location $location"
